Monthly Email: RELOADED
Mailer style has been directly pulled from the original mockup. Font is
still Arial as browsers aren’t fond of anything but the “web-safe
fonts”
(http://support.mailpoet.com/knowledgebase/which-fonts-can-i-use-html-ve
rsus-image/). Left it at Arial.

The newsletter style is now colored from the <html> tag instead of the
<body> tag. This /should/ give it color to the background. Obviously we
won’t know till a test email is sent since we can’t test that locally.

Thank-you.png is moved and renamed below so that its accessible facing
the web.

Button is added to the email. Hopefully it sticks.

Title is stylized as asked.