Hidden Footer & Tooltips
#scroll-fold and company added to stylesheet. Hides behind the body
content but is fixed. Margin is added to body element and it appears in
the void created. http://tjacobdesign.com/pure-css-hidden-footer/

Added hidden footer to footer.html.haml partial so it applies to every
page.

Uploaded logo images in hidden footer & tooltips with the content baked
right in.

World map uploaded as a background for the hidden footer.

Paypal link added to the checkout buttons. They now go their
appropriate places instead of holders.

———————

Tested across browsers and it seems to work flawlessly.

———————

BUG: there is a weird bug that seems to add a margin to the body
element or the wrapper. The only thing is though that it doesn’t show
up in any of the CSS and it also isn’t selectable in Inspector. Not
sure what this is. I ended up working it into the design because I
couldn’t make it go away. It is now the divider between the hidden
footer and the normal footer.