Signup Page & Monthly Email
Signup page isn’t linked to “Restaurants” anymore. It is its own page
now and restaurants is back to its original state. The code is still
there but its commented out.

Monthly mailer now has inline styling to force it to render correctly.
Not sure where its breaking along the pipe but this fixes it.