fb buttons & android bar
Buttons are aligned together when not using expanded view

Footer z-index changed so Facebook like banner thing should flow over
the content, not under it.

Android bar added and called. Not sure if it’ll work as the mobile
versions can’t be tested locally but there is no game breaking bugs
that take down the whole website, it’ll just not show up and I’ll have
to tweak it a little more.